+++ Source/EasyNetQ.DI.Tests/ResolverFactory.cs
@@ -1,0 +1,5 @@
+using System;
+
+namespace EasyNetQ.DI.Tests;
+
+public delegate IServiceResolver ResolverFactory(Action<IServiceRegister> configure);
